--- tests/test_compile_error.rs
@@ -1,5 +1,0 @@
-#[test]
-fn test_compile_errors_pyo3testing() {
-    let t = trybuild::TestCases::new();
-    t.compile_fail("tests/ui/invalid_pyo3imports.rs");
-}
