+++ tests/test_pyo3test.rs
@@ -1,5 +1,5 @@
-use pyo3::{exceptions::PyTypeError, prelude::*, types::PyDict};
-use pyo3_testing::{pyo3test, with_py_raises};
+use pyo3::{prelude::*, types::PyDict};
+use pyo3_testing::pyo3test;
 
 // The example from the Guide ...
 fn o3_addone(num: isize) -> isize {
@@ -154,20 +154,8 @@ fn test_star_args() {
     assert_eq!(result, 3)
 }
 
-#[pyo3test]
-#[allow(unused_macros)]
-#[pyo3import(py_adders: from adders import addone)]
-fn test_raises_validate_approach() {
-    match { addone.call1(("4",)) } {
-        Ok(_) => panic!("No Error"),
-        Err(error) if error.is_instance_of::<PyTypeError>(py) => return (),
-        Err(_) => panic!("Wrong Error"),
-    };
-}
-
-#[pyo3test]
-#[allow(unused_macros)]
-#[pyo3import(py_adders: from adders import addone)]
-fn test_raises() {
-    with_py_raises!(PyTypeError, { addone.call1(("4",)) }); //can't use `let result =` here
+#[test]
+fn test_compile_errors_pyo3testing() {
+    let t = trybuild::TestCases::new();
+    t.compile_fail("tests/ui/invalid_pyo3imports.rs");
 }
